Update VMImport custom resource
Update VMImport resource spec and status to make the ovirt source on
provider to be an optional, since we'd assume other sources will be
added in the future.

Also, changes are made to reduce the required levels of data and omit
the 'target' from the resource which didn't add a value.

Signed-off-by: Moti Asayag <masayag@redhat.com>