Add DiskMappings for specific disk mappings
DiskMappings will allow to map source VM disk to storage class.
A mapping for a disk overrides the mapping from  StorageMappings.

DiskMappings will fail to map the disk in the following cases:
* Disk ID/Name doesn't exist for the source VM (ID precedes name)
* Target storage class doesn't exist

DiskMappings.Source.Name should hold the disk alias.

DiskMappings.Source.ID should hold the ID of the disk and not the ID of
the disk-attachment that describes its association to the VM.

DiskMappings is optional, if not provided mappings will rely on
StorageMappings.

Signed-off-by: Moti Asayag <masayag@redhat.com>