Using common templates
Based on source vm os, workload we need to select a template to be
used for target vm. In case of guest os availble we map oses using
osInfo map. When we only have os - human redable form we attempt to
find which default os to use.

We use medium flavor since we use source vm to update the metadata
and it not important to pick amount of resources used in the template.

Signed-off-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>