Use the retrieved template name in vm label
The template processing adds the label 'vm.kubevirt.io/template' to
the vm with a value that is a trimmed version of the Template resource
name, i.e. rhel7-server-medium vs rhel7-server-medium-0.7.0.

By setting the real template name in the label, templates-admission is
capable to find the template and validate the VM.

Signed-off-by: Moti Asayag <masayag@redhat.com>