Set unique data-volume name
Setting data-volume name as target-vm-name + "-" + disk-attachment-id
will assure that the DV name is unique.
This will allow importing the same disk or the same vm with a different
target name.

Using 'GenerateName' option is problematic, since there is certain
logic that relies on a known and consistent DV Name prior to its
creation.

Signed-off-by: Moti Asayag <masayag@redhat.com>