Allow to import vms using i440fx bios
We required a user to make sure that source vm uses q35 compatibile bios.
Changing the bios do not modify the quest but it may affect device layout
upon reboot. We should not affect source vm while importing. This PR allows
to import i440fx in the same way as q35.

Signed-off-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>