Fix erroneous test failures caused by lingering resources
The "two imports of the same NIC-less source" oVirt test
was listing all VMs in order to check that there isn't a
duplicate in its namespace. This was causing it to fail
if a previous test's resources in a different namespace
had not been cleaned up properly. (in this case, a
vm left over from a resource mapping test)

This fixes the immediate issue by causing the above test
to only list VMs in its own namespace, but the cause
of resources not getting deleted correctly should still
be investigated.

Signed-off-by: Sam Lucidi <slucidi@redhat.com>