Skip the job and directly create a pod for guest conversion
The virt-v2v pod needs to be created by the vm-import-controller
in order for it to have access to the anyuid scc. If it is
created via a Job, the job-controller is responsible for
creating it and it will not have access to the needed scc.

This uses the anyuid scc to set the fsGroup for the virt-v2v
pod to the same one used by the cdi importer pod. This solves
a permissions problem that was preventing the virtio
overlay fom being commited properly when importing a VM's disk
to the nfs storage class.

Signed-off-by: Sam Lucidi <slucidi@redhat.com>