Requeue while waiting for disk copies and v2v
Imports are sometimes getting stuck waiting for disks
to copy or guest conversion to complete, even though
the datavolumes or v2v pod are done. Forcing the
VMI to reconcile (e.g. by manually adding an annotation)
allows the status change to be picked up and import to
complete successfully.

This PR causes VMIs to requeue themselves while waiting
for disk copy and guest conversion to be completed, to
prevent imports from getting stuck like this.

Signed-off-by: Sam Lucidi <slucidi@redhat.com>