Remove KubeVirt feature gate checks
KubeVirt LiveMigration being enabled or not has
no bearing on whether or not an oVirt VM can be
migrated to it, so this should not block import.

ImportWithoutTemplate is not a legitimate KubeVirt
feature gate, so it should be removed. It's unlikely
that anyone was actually using this as it was off
by default and not in the KubeVirt documentation.

Signed-off-by: Sam Lucidi <slucidi@redhat.com>